// src/utils/gmx/context/SyntheticsEvents/types.ts
export type PendingDepositData = {
  account: string;
  marketAddress: string;
  initialLongTokenAddress: string;
  initialShortTokenAddress: string;
  longTokenSwapPath: string[];
  shortTokenSwapPath: string[];
  initialLongTokenAmount: BigNumber;
  initialShortTokenAmount: BigNumber;
  minMarketTokens: BigNumber;
  shouldUnwrapNativeToken: boolean;
};

export type WithdrawalCreatedEventData = {
  key: string;
  account: string;
  receiver: string;
  callbackContract: string;
  marketAddress: string;
  marketTokenAmount: BigNumber;
  minLongTokenAmount: BigNumber;
  minShortTokenAmount: BigNumber;
  updatedAtBlock: BigNumber;
  executionFee: BigNumber;
  callbackGasLimit: BigNumber;
  shouldUnwrapNativeToken: boolean;
};

export type PendingWithdrawalData = {
  account: string;
  marketAddress: string;
  marketTokenAmount: BigNumber;
  minLongTokenAmount: BigNumber;
  minShortTokenAmount: BigNumber;
  shouldUnwrapNativeToken: boolean;
};